Licensing rework
The static methods weren't cutting it for me, because I somehow needed to change the way the types were filtered. I did something that resembles the decorator pattern in that a LicensedDispatcher both inherits from the abstract Dispatcher and wraps an implemented Dispatcher.

I also merged all the different projects into one because it was getting kinda unwieldy and I was starting to get the the point where I was needing bidirectional references between projects.

Co-Authored-By: Frederik Madsen <FredTheDane@users.noreply.github.com>