Updated the ValidatorDispatcher
Any dispatcher with a return type now also has a results cache
Added a speedtest for the ValidatorDispatcher to demonstrate.

Also added the three dispatchers as properties to the ConfigurableVaultApplication, so users don't have to worry about instantiation.
To make them LicensedDispatchers, users just add the [UseLicensing] attribute on the VaultApplication class